project/requirements.txt
```
  pytest
  pytest-html
  playwright
  Pillow
  imagehash
  
```